requirements.txt
django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
psycopg[binary]>=3.2,<3.3
django-cors-headers>=4.9,<5
python-decouple>=3.8,<4
drf-spectacular>=0.28,<0.29
